[COLAB-2918]-[comment]: rename package
