+++ microservices/services/comment-service/src/main/java/org/xcolab/service/comment/CommentServiceApplication.java
@@ -1,4 +1,4 @@
-package org.xcolab.service.comments;
+package org.xcolab.service.comment;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
