+++ microservices/services/comment-service/src/main/java/org/xcolab/service/comment/domain/categorygroup/CategoryGroupDao.java
@@ -1,7 +1,7 @@
-package org.xcolab.service.comments.domain.categorygroup;
+package org.xcolab.service.comment.domain.categorygroup;
 
 import org.xcolab.client.comment.pojo.ICategoryGroup;
-import org.xcolab.service.comments.exceptions.NotFoundException;
+import org.xcolab.service.comment.exceptions.NotFoundException;
 import org.xcolab.service.utils.PaginationHelper;
 
 import java.util.List;
